GeometryMerge Precision Fix
Currently the last remaining error is one of precision. Sometimes
PointInside(), LineInside(), PointOnLine(),  LineIntersectMuntac() all
fail to recognize a line that visually appears overlapping. It however
is just skirting a valid line and also not being picked up by
PointInside() or PointOnLine().

- So far the bandaid solution has been to go through all the edges of
the geometry and remove any edge that does not have a shared vertex with
another edge. For some reason this caused problems when it was
stipulated that each edge must have 2 partner edges it shares a vertex
with. (Possibly overlapping point which shares endpoint?)