Backup 15/4/15
Principle of two different methods that answer the same question is
being put in. Line intersection functions now have point on line and
counter-clockwise checks in addition to the previous method. There's
also a higher eps value used for comparisons in those functions. An
implementation bug in the ordering algorithm was fixed as well.