PointOutside() Fix
Previously, it was assumed that intersection of the ray with corners of
the polygon (i.e. shared endpoint or vertex) need to be ignored as they
would be adding a 2 to the intersection count.

But this fails in the most basic case where a point is inside the
polygon and the ray only intersects at the corner. The count isn't made
and an even count (in this case 0) tells the function that the point is
indeed outside.

Solution: Collect all the intersection points with the ray. These will
be unique. Put them in a list. The only intersection point that will
appear twice is the one found at the corner. Go through the list and for
each duplicate reduce the count by one.